Add dirtyness check to prevent unessecary hydration. Remove wrapper abstraction
