+++ src/System/Wrappers/ObjectWrapper.php
@@ -2,14 +2,36 @@
 
 namespace Analogue\ORM\System\Wrappers;
 
+use Analogue\ORM\EntityMap;
+use Analogue\ORM\System\InternallyMappable;
+use Analogue\ORM\System\Proxies\ProxyFactory;
 use Analogue\ORM\Exceptions\MappingException;
 use Zend\Hydrator\HydratorInterface;
 
 /**
  * Mixed wrapper using HydratorGenerator.
  */
-class ObjectWrapper extends Wrapper
+class ObjectWrapper implements InternallyMappable
 {
+    /**
+     * Original Entity Object.
+     *
+     * @var mixed
+     */
+    protected $entity;
+
+    /**
+     * Corresponding EntityMap.
+     *
+     * @var \Analogue\ORM\EntityMap
+     */
+    protected $entityMap;
+
+    /**
+     * @var \Analogue\ORM\System\Proxies\ProxyFactory
+     */
+    protected $proxyFactory;
+
     /**
      * Internal Representation of analogue's entity attributes.
      *
@@ -40,6 +62,14 @@ class ObjectWrapper extends Wrapper
     protected $hydrator;
 
     /**
+     * Set to true if the object's attributes have been modified since last
+     * hydration
+     * 
+     * @var boolean
+     */
+    protected $touched = false;
+
+    /**
      * Object Wrapper constructor.
      *
      * @param mixed                   $entity
@@ -49,10 +79,52 @@ class ObjectWrapper extends Wrapper
     public function __construct($entity, $entityMap, HydratorInterface $hydrator)
     {
         $this->hydrator = $hydrator;
-
-        parent::__construct($entity, $entityMap);
-
+        $this->entity = $entity;
+        $this->entityMap = $entityMap;
+        $this->proxyFactory = new ProxyFactory();
         $this->attributes = $this->dehydrate($entity);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getEntityClass(): string
+    {
+        return get_class($this->entity);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getEntityKeyName(): string
+    {
+        return $this->entityMap->getKeyName();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getEntityKeyValue()
+    {
+        return $this->getEntityAttribute($this->entityMap->getKeyName());
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getEntityHash(): string
+    {
+        return $this->getEntityClass().'.'.$this->getEntityKeyValue();
+    }
+
+    /**
+     * Returns the wrapped entity's map.
+     *
+     * @return mixed
+     */
+    public function getMap()
+    {
+        return $this->entityMap;
     }
 
     /**
@@ -62,7 +134,9 @@ class ObjectWrapper extends Wrapper
      */
     public function unwrap()
     {
-        $this->hydrate();
+        if($this->touched) {
+            $this->hydrate();
+        }
 
         return $this->entity;
     }
@@ -74,8 +148,6 @@ class ObjectWrapper extends Wrapper
      */
     public function getObject()
     {
-        //$this->hydrate();
-
         return $this->entity;
     }
 
@@ -115,6 +187,8 @@ class ObjectWrapper extends Wrapper
         }
 
         $this->hydrator->hydrate($properties, $this->entity);
+
+        $this->touched = false;
     }
 
     /**
@@ -200,8 +274,7 @@ class ObjectWrapper extends Wrapper
     public function setEntityAttributes(array $attributes)
     {
         $this->attributes = $attributes;
-
-        //$this->hydrate();
+        $this->touched = true;
     }
 
     /**
@@ -209,8 +282,6 @@ class ObjectWrapper extends Wrapper
      */
     public function getEntityAttributes(): array
     {
-        //$this->attributes = $this->dehydrate($this->entity);
-
         return $this->attributes;
     }
 
@@ -219,10 +290,10 @@ class ObjectWrapper extends Wrapper
      */
     public function setEntityAttribute(string $key, $value)
     {
-        //$this->attributes = $this->dehydrate($this->entity);
-
         $this->attributes[$key] = $value;
 
+        $this->touched = true;
+        
         $this->hydrate();
     }
 
@@ -231,8 +302,6 @@ class ObjectWrapper extends Wrapper
      */
     public function getEntityAttribute(string $key)
     {
-        //$this->attributes = $this->dehydrate($this->entity);
-
         if ($this->hasAttribute($key)) {
             return $this->attributes[$key];
         }
@@ -243,8 +312,125 @@ class ObjectWrapper extends Wrapper
      */
     public function hasAttribute(string $key): bool
     {
-        //$this->attributes = $this->dehydrate($this->entity);
-
         return array_key_exists($key, $this->attributes);
+    }
+
+    /**
+     * Set the lazy loading proxies on the wrapped entity objet.
+     *
+     * @param array $relations list of relations to be lazy loaded
+     *
+     * @return void
+     */
+    public function setProxies(array $relations = null)
+    {
+        $attributes = $this->getEntityAttributes();
+        $proxies = [];
+
+        $relations = $this->getRelationsToProxy();
+
+        // Before calling the relationship methods, we'll set the relationship
+        // method to null, to avoid hydration error on class properties
+        foreach ($relations as $relation) {
+            $this->setEntityAttribute($relation, null);
+        }
+
+        foreach ($relations as $relation) {
+
+            // First, we check that the relation has not been already
+            // set, in which case, we'll just pass.
+            if (array_key_exists($relation, $attributes) && !is_null($attributes[$relation])) {
+                continue;
+            }
+
+            // If the key is handled locally and we know it not to be set,
+            // we'll set the relationship to null value
+            if (!$this->relationNeedsProxy($relation, $attributes)) {
+                $proxies[$relation] = $this->entityMap->getEmptyValueForRelationship($relation);
+            } else {
+                $targetClass = $this->getClassToProxy($relation, $attributes);
+                $proxies[$relation] = $this->proxyFactory->make($this->getObject(), $relation, $targetClass);
+            }
+        }
+
+        foreach ($proxies as $key => $value) {
+            $this->setEntityAttribute($key, $value);
+        }
+    }
+
+    /**
+     * Get Target class to proxy for a one to one.
+     *
+     * @param string $relation
+     * @param array  $attributes
+     *
+     * @return string
+     */
+    protected function getClassToProxy($relation, array $attributes)
+    {
+        if ($this->entityMap->isPolymorphic($relation)) {
+            $localTypeAttribute = $this->entityMap->getLocalKeys($relation)['type'];
+
+            return $attributes[$localTypeAttribute];
+        }
+
+        return $this->entityMap->getTargettedClass($relation);
+    }
+
+    /**
+     * Determine which relations we have to build proxy for, by parsing
+     * attributes and finding methods that aren't set.
+     *
+     * @return array
+     */
+    protected function getRelationsToProxy()
+    {
+        $proxies = [];
+        $attributes = $this->getEntityAttributes();
+
+        foreach ($this->entityMap->getNonEmbeddedRelationships() as $relation) {
+            //foreach ($this->entityMap->getRelationships() as $relation) {
+
+            if (!array_key_exists($relation, $attributes) || is_null($attributes[$relation])) {
+                $proxies[] = $relation;
+            }
+        }
+
+        return $proxies;
+    }
+
+    /**
+     * Determine if the relation needs a proxy or not.
+     *
+     * @param string $relation
+     * @param array  $attributes
+     *
+     * @return bool
+     */
+    protected function relationNeedsProxy($relation, $attributes)
+    {
+        if (in_array($relation, $this->entityMap->getRelationshipsWithoutProxy())) {
+            return false;
+        }
+
+        $localKey = $this->entityMap->getLocalKeys($relation);
+
+        if (is_null($localKey)) {
+            return true;
+        }
+
+        if (is_array($localKey)) {
+            $localKey = $localKey['id'];
+        }
+
+        if (!isset($attributes[$localKey])) {
+            return false;
+        }
+
+        if (is_null($attributes[$localKey])) {
+            return false;
+        }
+
+        return true;
     }
 }
